requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
python-dotenv
httpx
